+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unrelease
 
 * maximum request length is 1024 bytes without, not with, CRLF
+* return 53 Proxy Request Refused for schemes different than gemini://
 
 ## 0.1.0